Keyboard repeat handling.
Some dialogs, the casebook for instance, are far easier handled by the keyboard than the mouse. To be really effective the cursor keys should be able to be repeated. This patch adds C and Lua code to enable keyboard repeat handling for any dialogs that need it.

Hotkey: C is for Casebook.

git-svn-id: http://corsix-th.googlecode.com/svn/trunk@399 c39591fa-788f-11de-a72b-d90af8dea425