Fix: The edit room button is now more responsive. Fix: Rooms should be registered correctly with the world after editing and aborting a room. (Issue 446) Fix: No difficulty dialog is shown if using the demo files. Fix: If easy or hard difficulty files are not found, first try to revert to normal difficulty. (Issue 453) Fix: Don't start the opening sequence until the tutorial is finished. Fix: Don't let reputation become a float after the autopsy has been discovered.
Added: The well known "It's far too cold" warning in the progress dialog.
Updated config templates with danish, the swedish language with strings for the lua console, and finnish according to issue 389. (thanks Jukka!)

WindowsInstaller: Don't delete old save games and config.txt when installing a new version on top of the old one. Don't show the second directory dialog if a config file is already present. Ask if the user wants to keep saved games when uninstalling. 

git-svn-id: http://corsix-th.googlecode.com/svn/trunk@998 c39591fa-788f-11de-a72b-d90af8dea425