New animations: some patients can fall over and some can shake there fist in anger when they have fallen (these are for when there are earth quakes in the game, but for now you may see them before then if you are lucky). Some patients might yawn if they are kept waiting.
Tweaked the diag process some more so that a consultant or other high skilled Dr can diagnose in one attempt.

Corrected what is believed to be a small spelling mistake in the build dialog from "payed" to "paid".

Changed the room name to the longer string for all rooms for the messages that show when you place staff in the wrong rooms (the adviser message should now make more sense), also now a message if you place a Dr or Nurse in the toilets.

Small change to the adviser messages about seating so that you are not told too often about needing benches.


Thanks, MarkL!

git-svn-id: http://corsix-th.googlecode.com/svn/trunk@1448 c39591fa-788f-11de-a72b-d90af8dea425