Modifies how handyman jobs are selected instead of plants and machines searching for handymen when used they are added in a task table from where handymen select their tasks.
Handymen priorities should be respected now (the sliding bars are now active).
If set to lowest, he will never do that task, and if set to highest, he will only do that task.

Handymen can also take over other handymen tasks. For example, if you have a handyman that goes to water a plant in the other side of the hospital and later a handyman near that plant decides to water it he will water that plant and the first handyman will search for another task to do.

Unfortunately there will be problems if you use this patch with existing saved games. For instance existing litter will never be cleaned and there may be more serious bugs, so start a fresh game after applying this revision.


Many thanks to alexghir9 for this patch!

git-svn-id: http://corsix-th.googlecode.com/svn/trunk@1588 c39591fa-788f-11de-a72b-d90af8dea425