Fix: Don't throw an unrecoverable error just because the TH install directory is wrong. Show the directory browser instead. Fix: Save config changes immediately. Change: Only show the "this is your first message" advisor message if playing the campaign. Players playing custom maps probably know about that already...
Added: The first-time-for-pay-raise advisor message.
Added: User actions are no longer allowed by default (for new games) when the game is paused. When trying this it might seem like it doesn't work correctly, but I've tried to mimic the original as much as possible. This for example means that a staff member being placed as you pause will disappear from the cursor. We might want to change the behaviour in the future though. Fixes issue 164.

git-svn-id: http://corsix-th.googlecode.com/svn/trunk@1786 c39591fa-788f-11de-a72b-d90af8dea425