Temporary: Since pausing doesn't work that well I have temporarily disabled it in code. The feature isn't that crucial anyway...
Fix: The adviser would not idle again if a new message arrived as he was idling from a previous message. Fixes issue 1513.
Fix: Don't play any announcements if there is receptionist to do it. Fixes issue 1517.
Fix: Since different languages contain different sets of "background noise" such as coughs, check each sound for existence before trying to play it. Related to issue 1431.

Added: Strings to the right of each line in the graph dialog showing the current value of each statistic. No additional strings have been added, so no need for additional localization because of this.

git-svn-id: http://corsix-th.googlecode.com/svn/trunk@2055 c39591fa-788f-11de-a72b-d90af8dea425